+++ libs/dana/src/components/data-display/index.ts
@@ -1,3 +1,2 @@
-export * from './card';
 export * from './table';
 export * from './divider';
